feat:easy locations and number format added

+++ lib/core/utils/get_earning_calculation.dart
@@ -1,4 +1,3 @@
-
-double getEarningCalculation(double money, int adet, int maliyet) {
+double getEarningCalculation(double money, int adet, double maliyet) {
   return (money * adet) - (maliyet * adet);
 }
